Remove infinitely waiting loop
Instead of waiting infinitely for the connection to be established, time out after 120s.

For testing purposes, don't quit after 120s and only print an error messages.

Perhaps the command used for testing whether the connection is established is not (or no longer) correct. Unfortunately I have no way of testing this locally.